// src/generatedTypes/IntrospectionQuery.ts
export interface IntrospectionQuery___schema_queryType {
  __typename: '__Type';
  name: string | null;
}

export interface IntrospectionQuery___schema_mutationType {
  __typename: '__Type';
  name: string | null;
}

export interface IntrospectionQuery___schema_subscriptionType {
  __typename: '__Type';
  name: string | null;
}
// ...
export interface IntrospectionQuery___schema_types {
  __typename: '__Type';
  kind: __TypeKind;
  name: string | null;
  description: string | null;
  fields: IntrospectionQuery___schema_types_fields[] | null;
  inputFields: IntrospectionQuery___schema_types_inputFields[] | null;
  interfaces: IntrospectionQuery___schema_types_interfaces[] | null;
  enumValues: IntrospectionQuery___schema_types_enumValues[] | null;
  possibleTypes: IntrospectionQuery___schema_types_possibleTypes[] | null;
}
// ...
export interface IntrospectionQuery___schema_directives {
  __typename: '__Directive';
  name: string;
  description: string | null;
  locations: __DirectiveLocation[];
  args: IntrospectionQuery___schema_directives_args[];
}

export interface IntrospectionQuery___schema {
  __typename: '__Schema';
  queryType: IntrospectionQuery___schema_queryType;
  mutationType: IntrospectionQuery___schema_mutationType | null;
  subscriptionType: IntrospectionQuery___schema_subscriptionType | null;
  types: IntrospectionQuery___schema_types[];
  directives: IntrospectionQuery___schema_directives[];
}

export interface IntrospectionQuery {
  __schema: IntrospectionQuery___schema;
}